Remove `ìnverse_of` from order has_many :adjustments
That option will cause trouble once Spree upgrades to Rails v4.0.0 as
the store will not create proper shipment adjustments. It appears that
`inverse_of: source` will confuse activerecord because an adjustment
source is not always an order. It might be either an Order or a
Shipment for example.

This is the confusing behaviour on rails 4 console session:

    [9] pry(Spree):1> order
    => #<Spree::Order id: 13, number: "R085433244" ... >
    [10] pry(Spree):1> shipment
    => #<Spree::Shipment id: 16, tracking: nil, number: "H33417384040" ... >
    [11] pry(Spree):1> order.adjustments.create(source: shipment, label: "ha", amount: 0)
    => #<Spree::Adjustment id: 41, source_id: 13, source_type: "Spree::Shipment", adjustable_id: 13, adjustable_type: "Spree::Order" ... >

While it also seems like an activerecord bug as well I still believe we
shouldn't rely on `inverse_of` where possible

Conflicts:
	core/app/models/spree/order.rb