Update _delivery.html.erb to enforce space between shipping method name and price
My `1-3-stable` site was showing the Shipping Method name and price running together - for some reason the space here wasn't registering. Converting it to a non-breaking space seemed to work, and so just to be certain about the desired behavior, I'd propose making it `&nbsp;` in Spree itself, in case there are other people encountering the same issue.

Fixes #3941