Show a big warning when people use protected_attributes
A lot of people have been running into issues when protected_attributes is in use, mainly because it FORCES models to have attr_accessible declared on them. This is not the default within Rails 4 engines/applications, and so we run into major problems with Spree when this gem is in use.