[api] Support order routes on /api/checkouts as well as /api/orders
I got sick of not knowing if /api/checkouts was the right prefix to use, or if I should use /api/orders. If I get sick of it after one day of usage, then I'd suspect other people who are using the API would also be sick of it.

Let's fix this problem and there will be no more getting sick of it :)