WIP: New Admin Order Interface
The admin order interface in Spree is one of the most painfully slow parts of Spree, and requires a lot of clicking in order to complete an order.

This is an attempt to fix that by using Backbone to display and manipulate the order's information by talking to Spree's API. So far, it looks like this has sped things up slightly.

There's still quite a lot of work to be done.